requirements.txt
absl-py==0.9.0
astor==0.8.1
charset-normalizer==3.3.2
cmake==3.28.1
docutils==0.20.1
face-recognition-models==0.3.0
facenet-pytorch==2.5.3
gast==0.2.2
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==3.6
importlib-metadata==1.7.0
imutils==0.5.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.0
opencv-python==4.3.0.36
opt-einsum==3.2.1
Pillow==9.5.0
protobuf==3.12.2
Pygments==2.17.2
pypiwin32==223
pywin32==306
PyYAML==5.3.1
requests==2.31.0
scipy==1.5.1
six==1.15.0
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
